#167  - Fix menu
