+++ run_detectia.sh
@@ -1,6 +1,4 @@
 #!/bin/bash
-source tflite2-env/bin/activate
+source /home/$USER/detectia-embebido/detectia-env/bin/activate
 sudo gpsd /dev/ttyUSB0 -F /var/run/gpsd.sock
-python3 ./script_deteccion/TFLite_detection_webcam.py --verbose
-
-
+python3 /home/$USER/detectia-embebido/script_deteccion/TFLite_menu_detection_webcam.py --verbose
